echo-os-tool/requirements.txt
moviepy==1.0.3
Pillow==10.4.0
numpy>=1.26.0
imageio>=2.34.0
tqdm>=4.66.0
